Fixed code structure of Userlogic
i have fixed the code structure of Userlogic (MVC)
